Cleanup PID file when the start command is terminated
When sigterm and friends are recieved, the server gets stopped but leaves behind
the PID file which causes future runs to think there is already a running
agent. With the recent adjustments to how termination events are handled, we can
safely cleanup that file in the stop method.

This PID file may be unneccessary, but this will at least ensure it is always
cleaned up until we have a better solution.